+++ cmp/tab/TabContainerModel.js
@@ -4,9 +4,11 @@
  *
  * Copyright © 2020 Extremely Heavy Industries Inc.
  */
+import {isEmpty} from 'lodash';
 import {HoistModel, managed, PersistenceProvider, RefreshMode, RenderMode, XH} from '@xh/hoist/core';
+import {div} from '../layout';
 import {action, observable} from '@xh/hoist/mobx';
-import {ensureNotEmpty, ensureUniqueBy, throwIf} from '@xh/hoist/utils/js';
+import {ensureUniqueBy, throwIf} from '@xh/hoist/utils/js';
 import {find, isUndefined} from 'lodash';
 import {TabModel} from './TabModel';
 
@@ -42,13 +44,17 @@ export class TabContainerModel {
     /** @member {RefreshMode} */
     refreshMode;
 
+    /** @member {(string|ReactNode)} */
+    emptyText;
+
     /**
      * @param {Object} c - TabContainerModel configuration.
      * @param {Object[]} c.tabs - configs for TabModels to be displayed.
      * @param {?string} [c.defaultTabId] - ID of Tab to be shown initially if routing does not
      *      specify otherwise. If not set, will default to first tab in the provided collection.
      * @param {?string} [c.route] - base route name for this container. If set, this container will
-     *      be route-enabled, with the route for each tab being "[route]/[tab.id]".  Cannot be used with `persistWith`.
+     *      be route-enabled, with the route for each tab being "[route]/[tab.id]".
+     *      Cannot be used with `persistWith`.
      * @param {string} [c.switcherPosition] - Position of the switcher docked within this component.
      *      Valid values are 'top', 'bottom', 'left', 'right', or 'none' if no switcher shown.
      * @param {boolean} [c.track] - True to enable activity tracking of tab views (default false).
@@ -56,7 +62,10 @@ export class TabContainerModel {
      *      per-tab via `TabModel.renderMode`. See enum for description of supported modes.
      * @param {RefreshMode} [c.refreshMode] - strategy for refreshing child tabs. Can be set
      *      per-tab via `TabModel.refreshMode`. See enum for description of supported modes.
-     * @param {PersistOptions} [c.persistWith] - options governing persistence.  Cannot be used with `route`.
+     * @param {PersistOptions} [c.persistWith] - options governing persistence.
+     *      Cannot be used with `route`.
+     * @param {(string|ReactNode)} [c.emptyText] - placeholder to display if no tabs are provided
+     *      or all tabs have been removed via their `omit` config.
      */
     constructor({
         tabs,
@@ -66,11 +75,21 @@ export class TabContainerModel {
         track = false,
         renderMode = RenderMode.LAZY,
         refreshMode = RefreshMode.ON_SHOW_LAZY,
-        persistWith = null
+        persistWith,
+        emptyText = 'No tabs to display.'
     }) {
 
         tabs = tabs.filter(p => !p.omit);
-        ensureNotEmpty(tabs, 'TabContainerModel needs at least one child.');
+        if (isEmpty(tabs)) {
+            tabs.push({
+                content: () => div({
+                    className: 'xh-text-color-accent xh-pad',
+                    item: emptyText
+                })
+            });
+            switcherPosition = 'none';
+        }
+
         ensureUniqueBy(tabs, 'id', 'Multiple TabContainerModel tabs have the same id.');
         throwIf(!['top', 'bottom', 'left', 'right', 'none'].includes(switcherPosition), 'Unsupported value for switcherPosition.');
         throwIf(route && persistWith, '"persistWith" and "route" cannot both be specified.');